#18 : автовыбор Person при логине на сайт.
