+++ EdSurvey/clients/forms.py
@@ -1,0 +1,8 @@
+# clients.forms
+# from django.contrib.auth.forms import AuthenticationForm
+# from django import forms
+# from django.core.exceptions import ValidationError
+
+
+# class LoginForm(AuthenticationForm):
+#     person = forms.CharField(label='aka', max_length=15)  # TODO not_empty
